+++ CHANGELOG.md
@@ -5,8 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.6] - 2025-06-18
 
-## [Unreleased]
+### Changed
+- Add `transport?: never;` to `CommandBasedConfig`
+- Remove mistakenly added `ws` from the dependencies
+
+## [0.2.5] - 2025-06-18
 
 ### Added
 - Issue #33: Streamable HTTP support